added event handling framework, file struct improvement
